refactor(store): export preconfigured store, clean up index
