+++ src/client/index.jsx
@@ -1,5 +1,6 @@
 import 'babel-polyfill';
 import '../styles/index';
+import './vendor/polyfills';
 import './vendor/nanoscroller';
 
 import React from 'react';
@@ -7,26 +8,12 @@ import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 
 import AppContainer from './pages/App';
-import configureStore from './store';
-import { loadState, saveState } from './store/localStorage';
-import { appReady } from './actions'
 
-const isProd = process.env.NODE_ENV === "production"
-const preloadedState = isProd ? loadState() : undefined;
-const store = configureStore(preloadedState);
+import store from './redux/store'
+import { appReady } from './redux/actions'
 
-// Save changes to localstorage each time an action is dispatched
-store.subscribe( () => {
-	saveState(store.getState());
-})
-
+// Setup trigger for preloader on app ready
 window.AppReady = () => store.dispatch(appReady())
-
-// Enable checking the store at any point 
-if (!isProd) 
-    window.getState = store.getState;
-
-console.info('Initial store:', store.getState());
 
 ReactDOM.render(
     <Provider store={store}>
